Information about the climate change data.

getInfoClimate returns information about the climate change data accessed from CPTEC/INPE.

getInfoClimate returns information about the climate change data accessed from CPTEC/INPE.

getInfoClimate()

getInfoClimate()

Value

Model driven, frequencies and variables.

Model driven, frequencies and variables.

Examples

getInfoClimate()

#> Available Models: o modelo (modelID) deve ser acessado pelo valor
#> do campo <id>. Verifique o respectivo periodo de abrangencia do modelo:
#> data inicial (iMonth, iYear) e data final (fMonth, fYear)
#> 
#>  id name                                        model couple     scenario  
#>  1  20 km, RCP4.5, continental, MIROC5.         ETA   MIROC5     RCP4.5    
#>  2  20 km, RCP8.5, continental, MIROC5.         ETA   MIROC5     RCP8.5    
#>  15 20 km, Histórico, continental, MIROC5.      ETA   MIROC5     Historical
#>  4  05 km, RCP4.5, sudeste-BR, HADGEM2-ES.      ETA   HADGEM2-ES RCP4.5    
#>  6  05 km, RCP8.5, sudeste-BR, HADGEM2-ES.      ETA   HADGEM2-ES RCP8.5    
#>  16 05 km, Histórico, sudeste-BR, HADGEM2-ES.   ETA   HADGEM2-ES Historical
#>  3  20 km, RCP4.5, continental, HADGEM2-ES.     ETA   HADGEM2-ES RCP4.5    
#>  5  20 km, RCP8.5, continental, HADGEM2-ES.     ETA   HADGEM2-ES RCP8.5    
#>  14 20 km, Histórico, continental, HADGEM2-ES.  ETA   HADGEM2-ES Historical
#>  7  20 km, RCP4.5, continental, CANESM2.        ETA   CANESM2    RCP4.5    
#>  8  20 km, RCP8.5, continental, CANESM2.        ETA   CANESM2    RCP8.5    
#>  13 20 km, Histórico, continental, CANESM2.     ETA   CANESM2    Historical
#>  20 20 km, RCP4.5, continental, BESM.           ETA   BESM       RCP4.5    
#>  19 20 km, RCP8.5, continental, BESM.           ETA   BESM       RCP8.5    
#>  17 20 km, Histórico, continental, BESM.        ETA   BESM       Historical
#>  22 05 km, RCP4.5, sudesteD2-BR, HADGEM2-ES.    ETA   HADGEM2-ES RCP4.5    
#>  23 05 km, RCP8.5, sudesteD2-BR, HADGEM2-ES.    ETA   HADGEM2-ES RCP8.5    
#>  21 05 km, Histórico, sudesteD2-BR, HADGEM2-ES. ETA   HADGEM2-ES Historical
#>  resolution start_year end_year
#>  20         2006       2099    
#>  20         2006       2099    
#>  20         1961       2005    
#>  5          2006       2099    
#>  5          2006       2099    
#>  5          1961       2005    
#>  20         2006       2099    
#>  20         2006       2099    
#>  20         1961       2005    
#>  20         2006       2099    
#>  20         2006       2099    
#>  20         1961       2005    
#>  20         2006       2099    
#>  20         2006       2099    
#>  20         1961       2005    
#>  5          2006       2099    
#>  5          2006       2099    
#>  5          1961       2005    
#> 
#> modelFrequency - Available frequencies:
#> HOURLY : horaria de 3 em 3 horas 
#> DAILY  : diaria
#> MONTHLY: mensal
#> YEARLY : anual
#> 
#> Available Variables - a variavel (modelVar) deve ser acessada
#> pelo seu short name <variable>
#>  variable
#>  CAPE    
#>  CDD     
#>  CDDc    
#>  CLSF    
#>  CSDI    
#>  CSDIc   
#>  CSSF    
#>  CWD     
#>  CWDc    
#>  D100    
#>  D10M    
#>  DP2M    
#>  DTR     
#>  DTRc    
#>  EVTP    
#>  FD      
#>  FDc     
#>  GHFL    
#>  HINV    
#>  ID      
#>  IDc     
#>  LWNV    
#>  MDNV    
#>  MNTP    
#>  MXTP    
#>  NEVE    
#>  OC10pc  
#>  OCES    
#>  OCIS    
#>  OCnc    
#>  OLES    
#>  OLIS    
#>  PRCPTOT 
#>  PRCPTOTc
#>  PREC    
#>  PS10pc  
#>  PS90pc  
#>  PSLC    
#>  PSLM    
#>  PSnc    
#>  PSxc    
#>  R10mm   
#>  R10mmc  
#>  R20mm   
#>  R20mmc  
#>  R25mm   
#>  R25mmc  
#>  R95p    
#>  R95pc   
#>  R99p    
#>  R99pc   
#>  RNOF    
#>  RX1day  
#>  RX1dayc 
#>  RX5day  
#>  RX5dayc 
#>  SDII    
#>  SDIIc   
#>  SU25    
#>  SU25c   
#>  TGRZ    
#>  TGSC    
#>  TN10p   
#>  TN10pc  
#>  TN90p   
#>  TN90pc  
#>  TNn     
#>  TNnc    
#>  TNx     
#>  TNxc    
#>  TP2M    
#>  TR      
#>  TRc     
#>  TSFC    
#>  TX10p   
#>  TX10pc  
#>  TX90p   
#>  TX90pc  
#>  TXn     
#>  TXnc    
#>  TXx     
#>  TXxc    
#>  UR2M    
#>  USSL    
#>  UZRS    
#>  W100    
#>  W10M    
#>  WSDI    
#>  WSDIc   
#>  WX10Pc  
#>  WX90pc  
#>  WXnc    
#>  WXxc    
#>  description                                                                                                 
#>  Energia Potencial Disponível para Convecção                                                                 
#>  Número máximo de dias consecutivos sem chuva no ano                                                         
#>  Número máximo de dias consecutivos sem chuva no ano - corrigido                                             
#>  Fluxo de Calor Latente à Superfície                                                                         
#>  Número máximo de dias consecutivos no ano com temperatura mínima diária menor que percentil 10              
#>  Número máximo de dias consecutivos no ano com temperatura mínima diária menor que percentil 10 - corrigido  
#>  Fluxo de Calor Sensível à Superfície                                                                        
#>  Número máximo de dias consecutivos com chuva no ano                                                         
#>  Número máximo de dias consecutivos com chuva no ano - corrigido                                             
#>  Vento a 100 metros - Direção meteorológica                                                                  
#>  Vento a 10 metros - Direção meteorológica                                                                   
#>  Temperatura do Ponto de Orvalho a 2 m                                                                       
#>  Amplitude anual média da temperatura diurna                                                                 
#>  Amplitude anual média da temperatura diurna - corrigido                                                     
#>  Evaporação                                                                                                  
#>  Total de dias com temperatura abaixo de 0ºC                                                                 
#>  Total de dias com temperatura abaixo de 0ºC - corrigido                                                     
#>  Fluxo de Calor no Solo                                                                                      
#>  Nebulosidade - Fração de cobertura de nuvens altas                                                          
#>  Total de dias com temperatura máxima abaixo de 0ºC                                                          
#>  Total de dias com temperatura máxima abaixo de 0ºC - corrigido                                              
#>  Nebulosidade - Fração de cobertura de nuvens baixas                                                         
#>  Nebulosidade - Fração de cobertura de nuvens médias                                                         
#>  Temperatura Mínima do ar a 2m                                                                               
#>  Temperatura Máxima do ar a 2m                                                                               
#>  Neve                                                                                                        
#>  Porcentagem anual de dias em que a radiação de onda curta incidente foi inferior ao percentil 10 - corrigido
#>  Radiação de Onda Curta Refletida à Superfície                                                               
#>  Radiação de Onda Curta Incidente à Superfície                                                               
#>  Menor valor da radiação de onda curta incidente anual - corrigido                                           
#>  Radiação de Onda Longa Refletida à Superfície                                                               
#>  Radiação de Onda Longa Incidente à Superfície                                                               
#>  Precipitação total anual                                                                                    
#>  Precipitação  total anual - corrigido                                                                       
#>  Precipitação Total                                                                                          
#>  Porcentagem anual de dias em que a pressão à superfície foi inferior ao percentil 10- corrigido             
#>  Porcentagem anual de dias em que a pressão à superfície excedeu o percentil 90 - corrigido                  
#>  Pressão à Superfície                                                                                        
#>  Pressão ao Nível Médio do Mar                                                                               
#>  Menor valor da pressão à superfície anual - corrigido                                                       
#>  Maior valor da pressão à superfície anual - corrigido                                                       
#>  Número de dias no ano com chuva acima de 10 mm                                                              
#>  Número de dias no ano com chuva acima de 10 mm - corrigido                                                  
#>  Número de dias no ano com chuva acima de 20 mm                                                              
#>  Número de dias no ano com chuva acima de 20 mm - corrigido                                                  
#>  Número de dias no ano com chuva acima de 25 mm                                                              
#>  Número de dias no ano com chuva acima de 25 mm - corrigido                                                  
#>  Precipitação total anual dos dias em que a chuva excedeu o percentil 95                                     
#>  Precipitação total anual dos dias em que a chuva excedeu o percentil 95 - corrigido                         
#>  Precipitação anual total dos dias em que a chuva excedeu o percentil 99                                     
#>  Precipitação anual total dos dias em que a chuva excedeu o percentil 99 - corrigido                         
#>  Escoamento Superficial                                                                                      
#>  Máxima precipitação anual em 1 dia                                                                          
#>  Máxima precipitação anual em 1 dia - corrigido                                                              
#>  Máxima precipitação anual em acumulada 5 dias consecutivos                                                  
#>  Máxima precipitação anual em acumulada 5 dias consecutivos - corrigido                                      
#>  Precipitação anual dividida pelo número de dias com chuva                                                   
#>  Precipitação anual dividida pelo número de dias com chuva - corrigido                                       
#>  Número de dias no ano em que a temperatura máxima excedeu 25ºC                                              
#>  Número de dias no ano em que a temperatura máxima excedeu 25°C - corrigido                                  
#>  Temperatura do Solo na Camada de 10-40 cm                                                                   
#>  Temperatura do Solo na Camada de 0-10 cm                                                                    
#>  Porcentagem anual de dias em que a temperatura mínima foi inferior ao percentil 10                          
#>  Porcentagem anual de dias em que a temperatura mínima foi inferior ao percentil 10 - corrigido              
#>  Porcentagem anual de dias em que a temperatura mínima excedeu o percentil 90                                
#>  Porcentagem anual de dias em que a temperatura mínima excedeu o percentil 90 - corrigido                    
#>  Menor temperatura mínima anual                                                                              
#>  Menor temperatura mínima anual - corrigido                                                                  
#>  Maior temperatura mínima anual                                                                              
#>  Maior temperatura mínima anual - corrigido                                                                  
#>  Temperatura do ar a 2 m                                                                                     
#>  Número de dias no ano em que a temperatura mínima excedeu 20°C                                              
#>  Número de dias no ano em que a temperatura mínima excedeu 20°C - corrigido                                  
#>  Temperatura Superficial (skin)                                                                              
#>  Porcentagem anual de dias em que a temperatura máxima foi inferior ao percentil 10                          
#>  Porcentagem anual de dias em que a temperatura máxima foi inferior ao percentil 10 - corrigido              
#>  Porcentagem anual de dias em que a temperatura máxima excedeu o percentil 90                                
#>  Porcentagem anual de dias em que a temperatura máxima excedeu o percentil 90 - corrigido                    
#>  Menor temperatura máxima anual                                                                              
#>  Menor temperatura máxima anual - corrigido                                                                  
#>  Maior temperatura máxima anual                                                                              
#>  Maior temperatura máxima anual - corrigido                                                                  
#>  Umidade Relativa do ar a 2m                                                                                 
#>  Umidade do Solo na Camada de 0-10 cm                                                                        
#>  Umidade do Solo na Camada de 10-40 cm                                                                       
#>  Vento a 100 metros - Intensidade                                                                            
#>  Vento a 10 metros - Intensidade                                                                             
#>  Número máximo de dias consecutivos no ano em que a temperatura máxima excedeu percentil 90                  
#>  Número máximo de dias consecutivos no ano em que a temperatura máxima excedeu percentil 90 - corrigido      
#>  Porcentagem anual de dias em que o vento máximo foi inferior ao percentil 10 - corrigido                    
#>  Porcentagem anual de dias em que o vento máximo excedeu o percentil 90- corrigido                           
#>  Menor intensidade do vento máximo anual - corrigido                                                         
#>  Maior intensidade do vento máximo anual - corrigido                                                         
#>  unit                
#>  J/kg                
#>  dias                
#>  dias                
#>  W/m^2               
#>  dias                
#>  dias                
#>  W/m^2               
#>  dias                
#>  dias                
#>  graus meteorológicos
#>  graus meteorológicos
#>  °C                  
#>  °C                  
#>  °C                  
#>  mm                  
#>  dias                
#>  dias                
#>  W/m^2               
#>  0-1                 
#>  dias                
#>  dias                
#>  0-1                 
#>  0-1                 
#>  °C                  
#>  °C                  
#>  mm                  
#>  %                   
#>  W/m^2               
#>  W/m^2               
#>  W/m^2               
#>  W/m^2               
#>  W/m^2               
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  hPa                 
#>  hPa                 
#>  hPa                 
#>  hPa                 
#>  dias                
#>  dias                
#>  dias                
#>  dias                
#>  dias                
#>  dias                
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  dias                
#>  dias                
#>  °C                  
#>  °C                  
#>  %                   
#>  %                   
#>  %                   
#>  %                   
#>  °C                  
#>  °C                  
#>  °C                  
#>  °C                  
#>  °C                  
#>  dias                
#>  dias                
#>  °C                  
#>  %                   
#>  %                   
#>  %                   
#>  %                   
#>  °C                  
#>  %                   
#>  °C                  
#>  %                   
#>  %                   
#>  0-1                 
#>  0-1                 
#>  m/s                 
#>  m/s                 
#>  dias                
#>  dias                
#>  %                   
#>  %                   
#>  m/s                 
#>  m/s                 

getInfoClimate()

#> Available Models: o modelo (modelID) deve ser acessado pelo valor
#> do campo <id>. Verifique o respectivo periodo de abrangencia do modelo:
#> data inicial (iMonth, iYear) e data final (fMonth, fYear)
#> 
#>  id name                                        model couple     scenario  
#>  1  20 km, RCP4.5, continental, MIROC5.         ETA   MIROC5     RCP4.5    
#>  2  20 km, RCP8.5, continental, MIROC5.         ETA   MIROC5     RCP8.5    
#>  15 20 km, Histórico, continental, MIROC5.      ETA   MIROC5     Historical
#>  4  05 km, RCP4.5, sudeste-BR, HADGEM2-ES.      ETA   HADGEM2-ES RCP4.5    
#>  6  05 km, RCP8.5, sudeste-BR, HADGEM2-ES.      ETA   HADGEM2-ES RCP8.5    
#>  16 05 km, Histórico, sudeste-BR, HADGEM2-ES.   ETA   HADGEM2-ES Historical
#>  3  20 km, RCP4.5, continental, HADGEM2-ES.     ETA   HADGEM2-ES RCP4.5    
#>  5  20 km, RCP8.5, continental, HADGEM2-ES.     ETA   HADGEM2-ES RCP8.5    
#>  14 20 km, Histórico, continental, HADGEM2-ES.  ETA   HADGEM2-ES Historical
#>  7  20 km, RCP4.5, continental, CANESM2.        ETA   CANESM2    RCP4.5    
#>  8  20 km, RCP8.5, continental, CANESM2.        ETA   CANESM2    RCP8.5    
#>  13 20 km, Histórico, continental, CANESM2.     ETA   CANESM2    Historical
#>  20 20 km, RCP4.5, continental, BESM.           ETA   BESM       RCP4.5    
#>  19 20 km, RCP8.5, continental, BESM.           ETA   BESM       RCP8.5    
#>  17 20 km, Histórico, continental, BESM.        ETA   BESM       Historical
#>  22 05 km, RCP4.5, sudesteD2-BR, HADGEM2-ES.    ETA   HADGEM2-ES RCP4.5    
#>  23 05 km, RCP8.5, sudesteD2-BR, HADGEM2-ES.    ETA   HADGEM2-ES RCP8.5    
#>  21 05 km, Histórico, sudesteD2-BR, HADGEM2-ES. ETA   HADGEM2-ES Historical
#>  resolution start_year end_year
#>  20         2006       2099    
#>  20         2006       2099    
#>  20         1961       2005    
#>  5          2006       2099    
#>  5          2006       2099    
#>  5          1961       2005    
#>  20         2006       2099    
#>  20         2006       2099    
#>  20         1961       2005    
#>  20         2006       2099    
#>  20         2006       2099    
#>  20         1961       2005    
#>  20         2006       2099    
#>  20         2006       2099    
#>  20         1961       2005    
#>  5          2006       2099    
#>  5          2006       2099    
#>  5          1961       2005    
#> 
#> modelFrequency - Available frequencies:
#> HOURLY : horaria de 3 em 3 horas 
#> DAILY  : diaria
#> MONTHLY: mensal
#> YEARLY : anual
#> 
#> Available Variables - a variavel (modelVar) deve ser acessada
#> pelo seu short name <variable>
#>  variable
#>  CAPE    
#>  CDD     
#>  CDDc    
#>  CLSF    
#>  CSDI    
#>  CSDIc   
#>  CSSF    
#>  CWD     
#>  CWDc    
#>  D100    
#>  D10M    
#>  DP2M    
#>  DTR     
#>  DTRc    
#>  EVTP    
#>  FD      
#>  FDc     
#>  GHFL    
#>  HINV    
#>  ID      
#>  IDc     
#>  LWNV    
#>  MDNV    
#>  MNTP    
#>  MXTP    
#>  NEVE    
#>  OC10pc  
#>  OCES    
#>  OCIS    
#>  OCnc    
#>  OLES    
#>  OLIS    
#>  PRCPTOT 
#>  PRCPTOTc
#>  PREC    
#>  PS10pc  
#>  PS90pc  
#>  PSLC    
#>  PSLM    
#>  PSnc    
#>  PSxc    
#>  R10mm   
#>  R10mmc  
#>  R20mm   
#>  R20mmc  
#>  R25mm   
#>  R25mmc  
#>  R95p    
#>  R95pc   
#>  R99p    
#>  R99pc   
#>  RNOF    
#>  RX1day  
#>  RX1dayc 
#>  RX5day  
#>  RX5dayc 
#>  SDII    
#>  SDIIc   
#>  SU25    
#>  SU25c   
#>  TGRZ    
#>  TGSC    
#>  TN10p   
#>  TN10pc  
#>  TN90p   
#>  TN90pc  
#>  TNn     
#>  TNnc    
#>  TNx     
#>  TNxc    
#>  TP2M    
#>  TR      
#>  TRc     
#>  TSFC    
#>  TX10p   
#>  TX10pc  
#>  TX90p   
#>  TX90pc  
#>  TXn     
#>  TXnc    
#>  TXx     
#>  TXxc    
#>  UR2M    
#>  USSL    
#>  UZRS    
#>  W100    
#>  W10M    
#>  WSDI    
#>  WSDIc   
#>  WX10Pc  
#>  WX90pc  
#>  WXnc    
#>  WXxc    
#>  description                                                                                                 
#>  Energia Potencial Disponível para Convecção                                                                 
#>  Número máximo de dias consecutivos sem chuva no ano                                                         
#>  Número máximo de dias consecutivos sem chuva no ano - corrigido                                             
#>  Fluxo de Calor Latente à Superfície                                                                         
#>  Número máximo de dias consecutivos no ano com temperatura mínima diária menor que percentil 10              
#>  Número máximo de dias consecutivos no ano com temperatura mínima diária menor que percentil 10 - corrigido  
#>  Fluxo de Calor Sensível à Superfície                                                                        
#>  Número máximo de dias consecutivos com chuva no ano                                                         
#>  Número máximo de dias consecutivos com chuva no ano - corrigido                                             
#>  Vento a 100 metros - Direção meteorológica                                                                  
#>  Vento a 10 metros - Direção meteorológica                                                                   
#>  Temperatura do Ponto de Orvalho a 2 m                                                                       
#>  Amplitude anual média da temperatura diurna                                                                 
#>  Amplitude anual média da temperatura diurna - corrigido                                                     
#>  Evaporação                                                                                                  
#>  Total de dias com temperatura abaixo de 0ºC                                                                 
#>  Total de dias com temperatura abaixo de 0ºC - corrigido                                                     
#>  Fluxo de Calor no Solo                                                                                      
#>  Nebulosidade - Fração de cobertura de nuvens altas                                                          
#>  Total de dias com temperatura máxima abaixo de 0ºC                                                          
#>  Total de dias com temperatura máxima abaixo de 0ºC - corrigido                                              
#>  Nebulosidade - Fração de cobertura de nuvens baixas                                                         
#>  Nebulosidade - Fração de cobertura de nuvens médias                                                         
#>  Temperatura Mínima do ar a 2m                                                                               
#>  Temperatura Máxima do ar a 2m                                                                               
#>  Neve                                                                                                        
#>  Porcentagem anual de dias em que a radiação de onda curta incidente foi inferior ao percentil 10 - corrigido
#>  Radiação de Onda Curta Refletida à Superfície                                                               
#>  Radiação de Onda Curta Incidente à Superfície                                                               
#>  Menor valor da radiação de onda curta incidente anual - corrigido                                           
#>  Radiação de Onda Longa Refletida à Superfície                                                               
#>  Radiação de Onda Longa Incidente à Superfície                                                               
#>  Precipitação total anual                                                                                    
#>  Precipitação  total anual - corrigido                                                                       
#>  Precipitação Total                                                                                          
#>  Porcentagem anual de dias em que a pressão à superfície foi inferior ao percentil 10- corrigido             
#>  Porcentagem anual de dias em que a pressão à superfície excedeu o percentil 90 - corrigido                  
#>  Pressão à Superfície                                                                                        
#>  Pressão ao Nível Médio do Mar                                                                               
#>  Menor valor da pressão à superfície anual - corrigido                                                       
#>  Maior valor da pressão à superfície anual - corrigido                                                       
#>  Número de dias no ano com chuva acima de 10 mm                                                              
#>  Número de dias no ano com chuva acima de 10 mm - corrigido                                                  
#>  Número de dias no ano com chuva acima de 20 mm                                                              
#>  Número de dias no ano com chuva acima de 20 mm - corrigido                                                  
#>  Número de dias no ano com chuva acima de 25 mm                                                              
#>  Número de dias no ano com chuva acima de 25 mm - corrigido                                                  
#>  Precipitação total anual dos dias em que a chuva excedeu o percentil 95                                     
#>  Precipitação total anual dos dias em que a chuva excedeu o percentil 95 - corrigido                         
#>  Precipitação anual total dos dias em que a chuva excedeu o percentil 99                                     
#>  Precipitação anual total dos dias em que a chuva excedeu o percentil 99 - corrigido                         
#>  Escoamento Superficial                                                                                      
#>  Máxima precipitação anual em 1 dia                                                                          
#>  Máxima precipitação anual em 1 dia - corrigido                                                              
#>  Máxima precipitação anual em acumulada 5 dias consecutivos                                                  
#>  Máxima precipitação anual em acumulada 5 dias consecutivos - corrigido                                      
#>  Precipitação anual dividida pelo número de dias com chuva                                                   
#>  Precipitação anual dividida pelo número de dias com chuva - corrigido                                       
#>  Número de dias no ano em que a temperatura máxima excedeu 25ºC                                              
#>  Número de dias no ano em que a temperatura máxima excedeu 25°C - corrigido                                  
#>  Temperatura do Solo na Camada de 10-40 cm                                                                   
#>  Temperatura do Solo na Camada de 0-10 cm                                                                    
#>  Porcentagem anual de dias em que a temperatura mínima foi inferior ao percentil 10                          
#>  Porcentagem anual de dias em que a temperatura mínima foi inferior ao percentil 10 - corrigido              
#>  Porcentagem anual de dias em que a temperatura mínima excedeu o percentil 90                                
#>  Porcentagem anual de dias em que a temperatura mínima excedeu o percentil 90 - corrigido                    
#>  Menor temperatura mínima anual                                                                              
#>  Menor temperatura mínima anual - corrigido                                                                  
#>  Maior temperatura mínima anual                                                                              
#>  Maior temperatura mínima anual - corrigido                                                                  
#>  Temperatura do ar a 2 m                                                                                     
#>  Número de dias no ano em que a temperatura mínima excedeu 20°C                                              
#>  Número de dias no ano em que a temperatura mínima excedeu 20°C - corrigido                                  
#>  Temperatura Superficial (skin)                                                                              
#>  Porcentagem anual de dias em que a temperatura máxima foi inferior ao percentil 10                          
#>  Porcentagem anual de dias em que a temperatura máxima foi inferior ao percentil 10 - corrigido              
#>  Porcentagem anual de dias em que a temperatura máxima excedeu o percentil 90                                
#>  Porcentagem anual de dias em que a temperatura máxima excedeu o percentil 90 - corrigido                    
#>  Menor temperatura máxima anual                                                                              
#>  Menor temperatura máxima anual - corrigido                                                                  
#>  Maior temperatura máxima anual                                                                              
#>  Maior temperatura máxima anual - corrigido                                                                  
#>  Umidade Relativa do ar a 2m                                                                                 
#>  Umidade do Solo na Camada de 0-10 cm                                                                        
#>  Umidade do Solo na Camada de 10-40 cm                                                                       
#>  Vento a 100 metros - Intensidade                                                                            
#>  Vento a 10 metros - Intensidade                                                                             
#>  Número máximo de dias consecutivos no ano em que a temperatura máxima excedeu percentil 90                  
#>  Número máximo de dias consecutivos no ano em que a temperatura máxima excedeu percentil 90 - corrigido      
#>  Porcentagem anual de dias em que o vento máximo foi inferior ao percentil 10 - corrigido                    
#>  Porcentagem anual de dias em que o vento máximo excedeu o percentil 90- corrigido                           
#>  Menor intensidade do vento máximo anual - corrigido                                                         
#>  Maior intensidade do vento máximo anual - corrigido                                                         
#>  unit                
#>  J/kg                
#>  dias                
#>  dias                
#>  W/m^2               
#>  dias                
#>  dias                
#>  W/m^2               
#>  dias                
#>  dias                
#>  graus meteorológicos
#>  graus meteorológicos
#>  °C                  
#>  °C                  
#>  °C                  
#>  mm                  
#>  dias                
#>  dias                
#>  W/m^2               
#>  0-1                 
#>  dias                
#>  dias                
#>  0-1                 
#>  0-1                 
#>  °C                  
#>  °C                  
#>  mm                  
#>  %                   
#>  W/m^2               
#>  W/m^2               
#>  W/m^2               
#>  W/m^2               
#>  W/m^2               
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  hPa                 
#>  hPa                 
#>  hPa                 
#>  hPa                 
#>  dias                
#>  dias                
#>  dias                
#>  dias                
#>  dias                
#>  dias                
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  mm                  
#>  dias                
#>  dias                
#>  °C                  
#>  °C                  
#>  %                   
#>  %                   
#>  %                   
#>  %                   
#>  °C                  
#>  °C                  
#>  °C                  
#>  °C                  
#>  °C                  
#>  dias                
#>  dias                
#>  °C                  
#>  %                   
#>  %                   
#>  %                   
#>  %                   
#>  °C                  
#>  %                   
#>  °C                  
#>  %                   
#>  %                   
#>  0-1                 
#>  0-1                 
#>  m/s                 
#>  m/s                 
#>  dias                
#>  dias                
#>  %                   
#>  %                   
#>  m/s                 
#>  m/s